requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
boto3==1.38.25
botocore==1.38.25
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
griffe==1.7.3
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.2
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
litellm==1.71.2
MarkupSafe==3.0.2
mcp==1.9.1
multidict==6.4.4
openai==1.82.0
openai-agents==0.0.16
packaging==25.0
propcache==0.3.1
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.46.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
zipp==3.22.0
